+++ client/src/components/pages/Store/Products/ProductsList/index.js
@@ -1,0 +1,1 @@
+export * from "./ProductsList";
